fix: 将DIM_LAST转换为int64_t类型以避免类型不匹配
修复NPU命令调用时的潜在类型不匹配问题，将DIM_LAST常量显式转换为int64_t类型。同时添加了cstdint头文件以确保int64_t类型的可用性。

+++ examples/inductor_pattern_match/torch_op_plugin/sub_mul_concat.cpp
@@ -13,6 +13,7 @@ See the License for the specific language governing permissions and
         limitations under the License.
 ==============================================================================*/
 
+#include <cstdint>
 #include "torch/csrc/autograd/custom_function.h"
 #include "torch/library.h"
 
@@ -29,7 +30,7 @@ at::Tensor SubMulConcat(const at::Tensor& lhs, const at::Tensor& rhs)
 {
     const at::OptionalDeviceGuard guard(device_of(lhs));
     auto output = at::empty({lhs.size(0), lhs.size(1), NUM_CONCAT * lhs.size(DIM_LAST)}, lhs.options());
-    EXEC_NPU_CMD(aclnnSubMulConcat, lhs, rhs, DIM_LAST, output);
+    EXEC_NPU_CMD(aclnnSubMulConcat, lhs, rhs, static_cast<int64_t>(DIM_LAST), output);
 }
 }  // namespace acl_ops
